+++ src/redi/cli/attachment_command.py
@@ -7,7 +7,7 @@ from redi.api.attachment import (
     update_attachment,
 )
 from redi.cli.alias import resolve_alias
-from redi.cli.prompt_util import confirm_delete
+from redi.cli.confirm import confirm_delete
 from redi.i18n import messages
