+++ src/redi/cli/issue_command.py
@@ -10,7 +10,7 @@ from prompt_toolkit.validation import Validator
 from redi.cli.alias import resolve_alias
 from redi.cli.editor import open_editor
 from redi.cli.picker import inline_checkbox, inline_choice
-from redi.cli.prompt_util import confirm_delete
+from redi.cli.confirm import confirm_delete
 from redi.cli.validator import (
     DateValidator,
     DueDateValidator,
