+++ src/redi/cli/time_entry_command.py
@@ -7,7 +7,7 @@ from prompt_toolkit.validation import Validator
 from redi.cli.alias import resolve_alias
 from redi.cli.keybinding import digit_and_period_key_bindings, digit_only_key_bindings
 from redi.cli.picker import inline_checkbox, inline_choice
-from redi.cli.prompt_util import confirm_delete
+from redi.cli.confirm import confirm_delete
 from redi.cli.validator import HourValidator
 from redi.config import default_project_id
 from redi.i18n import messages
