+++ src/redi/cli/version_command.py
@@ -9,7 +9,7 @@ from prompt_toolkit.validation import Validator
 
 from redi.cli.alias import resolve_alias
 from redi.cli.picker import inline_checkbox, inline_choice
-from redi.cli.prompt_util import confirm_delete
+from redi.cli.confirm import confirm_delete
 from redi.config import default_project_id
 from redi.i18n import messages
 from redi.api.version import (
