+++ src/redi/cli/wiki_command.py
@@ -6,7 +6,7 @@ from prompt_toolkit.validation import ValidationError, Validator
 from redi.cli.alias import resolve_alias
 from redi.cli.editor import open_editor
 from redi.cli.picker import inline_choice
-from redi.cli.prompt_util import confirm_delete
+from redi.cli.confirm import confirm_delete
 from redi.config import default_project_id, wiki_project_id
 from redi.i18n import messages
 from redi.api.wiki import (
