+++ tests/unit/cli/test_confirm.py
@@ -1,6 +1,6 @@
 import pytest
 
-from redi.cli import prompt_util
+from redi.cli import confirm
 from redi.i18n import messages
 
 
@@ -9,37 +9,37 @@ class TestConfirmDelete:
 
     def test_accepts_yes(self, monkeypatch, capsys):
         """'yes'の入力なら例外なく処理が続行する"""
-        monkeypatch.setattr(prompt_util, "prompt", lambda _msg: "yes")
-        prompt_util.confirm_delete("削除するX: 1")
+        monkeypatch.setattr(confirm, "prompt", lambda _msg: "yes")
+        confirm.confirm_delete("削除するX: 1")
         out = capsys.readouterr().out
         assert "削除するX: 1" in out
 
     def test_accepts_yes_case_insensitive(self, monkeypatch):
         """大文字小文字と前後空白を許容する"""
-        monkeypatch.setattr(prompt_util, "prompt", lambda _msg: "  YES  ")
-        prompt_util.confirm_delete("summary")
+        monkeypatch.setattr(confirm, "prompt", lambda _msg: "  YES  ")
+        confirm.confirm_delete("summary")
 
     def test_rejects_no(self, monkeypatch, capsys):
         """'no'ならexit(1)してキャンセルメッセージを出力する"""
-        monkeypatch.setattr(prompt_util, "prompt", lambda _msg: "no")
+        monkeypatch.setattr(confirm, "prompt", lambda _msg: "no")
         with pytest.raises(SystemExit) as exc:
-            prompt_util.confirm_delete("summary")
+            confirm.confirm_delete("summary")
         assert exc.value.code == 1
         assert messages.canceled in capsys.readouterr().out
 
     def test_rejects_empty_input(self, monkeypatch):
         """空入力はキャンセル扱い（デフォルトNoの挙動）"""
-        monkeypatch.setattr(prompt_util, "prompt", lambda _msg: "")
+        monkeypatch.setattr(confirm, "prompt", lambda _msg: "")
         with pytest.raises(SystemExit) as exc:
-            prompt_util.confirm_delete("summary")
+            confirm.confirm_delete("summary")
         assert exc.value.code == 1
 
     def test_rejects_other_inputs(self, monkeypatch):
         """'y'単体や関係ない文字列はキャンセル扱い"""
         for value in ["y", "ye", "n", "abc"]:
-            monkeypatch.setattr(prompt_util, "prompt", lambda _msg, v=value: v)
+            monkeypatch.setattr(confirm, "prompt", lambda _msg, v=value: v)
             with pytest.raises(SystemExit) as exc:
-                prompt_util.confirm_delete("summary")
+                confirm.confirm_delete("summary")
             assert exc.value.code == 1
 
     def test_keyboard_interrupt_cancels(self, monkeypatch, capsys):
@@ -48,9 +48,9 @@ class TestConfirmDelete:
         def raise_interrupt(_msg):
             raise KeyboardInterrupt
 
-        monkeypatch.setattr(prompt_util, "prompt", raise_interrupt)
+        monkeypatch.setattr(confirm, "prompt", raise_interrupt)
         with pytest.raises(SystemExit) as exc:
-            prompt_util.confirm_delete("summary")
+            confirm.confirm_delete("summary")
         assert exc.value.code == 1
         assert messages.canceled in capsys.readouterr().out
 
@@ -60,9 +60,9 @@ class TestConfirmDelete:
         def raise_eof(_msg):
             raise EOFError
 
-        monkeypatch.setattr(prompt_util, "prompt", raise_eof)
+        monkeypatch.setattr(confirm, "prompt", raise_eof)
         with pytest.raises(SystemExit) as exc:
-            prompt_util.confirm_delete("summary")
+            confirm.confirm_delete("summary")
         assert exc.value.code == 1
 
 
@@ -71,33 +71,33 @@ class TestConfirmDeleteWithIdentifier:
 
     def test_accepts_matching_identifier(self, monkeypatch, capsys):
         """識別子が一致すれば例外なく処理が続行する"""
-        monkeypatch.setattr(prompt_util, "prompt", lambda _msg: "my-project")
-        prompt_util.confirm_delete_with_identifier(
+        monkeypatch.setattr(confirm, "prompt", lambda _msg: "my-project")
+        confirm.confirm_delete_with_identifier(
             "削除するプロジェクト: 1 My Project", "my-project", "プロジェクト識別子"
         )
         assert "削除するプロジェクト: 1 My Project" in capsys.readouterr().out
 
     def test_trims_whitespace(self, monkeypatch):
         """前後空白は無視する"""
-        monkeypatch.setattr(prompt_util, "prompt", lambda _msg: "  my-project  ")
-        prompt_util.confirm_delete_with_identifier(
+        monkeypatch.setattr(confirm, "prompt", lambda _msg: "  my-project  ")
+        confirm.confirm_delete_with_identifier(
             "summary", "my-project", "プロジェクト識別子"
         )
 
     def test_is_case_sensitive(self, monkeypatch):
         """識別子の比較は大文字小文字を区別する"""
-        monkeypatch.setattr(prompt_util, "prompt", lambda _msg: "MY-PROJECT")
+        monkeypatch.setattr(confirm, "prompt", lambda _msg: "MY-PROJECT")
         with pytest.raises(SystemExit) as exc:
-            prompt_util.confirm_delete_with_identifier(
+            confirm.confirm_delete_with_identifier(
                 "summary", "my-project", "プロジェクト識別子"
             )
         assert exc.value.code == 1
 
     def test_rejects_mismatched_identifier(self, monkeypatch, capsys):
         """識別子が一致しなければexit(1)してメッセージを出力する"""
-        monkeypatch.setattr(prompt_util, "prompt", lambda _msg: "wrong-id")
+        monkeypatch.setattr(confirm, "prompt", lambda _msg: "wrong-id")
         with pytest.raises(SystemExit) as exc:
-            prompt_util.confirm_delete_with_identifier(
+            confirm.confirm_delete_with_identifier(
                 "summary", "my-project", "プロジェクト識別子"
             )
         assert exc.value.code == 1
@@ -108,9 +108,9 @@ class TestConfirmDeleteWithIdentifier:
 
     def test_rejects_empty_input(self, monkeypatch):
         """空入力は不一致扱い"""
-        monkeypatch.setattr(prompt_util, "prompt", lambda _msg: "")
+        monkeypatch.setattr(confirm, "prompt", lambda _msg: "")
         with pytest.raises(SystemExit) as exc:
-            prompt_util.confirm_delete_with_identifier(
+            confirm.confirm_delete_with_identifier(
                 "summary", "my-project", "プロジェクト識別子"
             )
         assert exc.value.code == 1
@@ -121,9 +121,9 @@ class TestConfirmDeleteWithIdentifier:
         def raise_interrupt(_msg):
             raise KeyboardInterrupt
 
-        monkeypatch.setattr(prompt_util, "prompt", raise_interrupt)
+        monkeypatch.setattr(confirm, "prompt", raise_interrupt)
         with pytest.raises(SystemExit) as exc:
-            prompt_util.confirm_delete_with_identifier(
+            confirm.confirm_delete_with_identifier(
                 "summary", "my-project", "プロジェクト識別子"
             )
         assert exc.value.code == 1
@@ -137,8 +137,8 @@ class TestConfirmDeleteWithIdentifier:
             captured["msg"] = msg
             return "my-project"
 
-        monkeypatch.setattr(prompt_util, "prompt", capture_prompt)
-        prompt_util.confirm_delete_with_identifier(
+        monkeypatch.setattr(confirm, "prompt", capture_prompt)
+        confirm.confirm_delete_with_identifier(
             "summary", "my-project", "プロジェクト識別子"
         )
         assert "プロジェクト識別子" in captured["msg"]
